Patching Network Settings

+++ src/networksettings/connman/qnetworksettingsmanager_p.cpp
@@ -181,20 +181,23 @@ void QNetworkSettingsManagerPrivate::onTechnologyAdded(const QDBusObjectPath &te
     Q_Q(QNetworkSettingsManager);
 
     foreach (QNetworkSettingsInterface* item, m_interfaceModel.getModel()) {
-        ConnmanSettingsInterface* tech = qobject_cast<ConnmanSettingsInterface*>(item);
-        if (tech->path() != technology.path()) {
-            ConnmanSettingsInterface *interface = new ConnmanSettingsInterface(technology.path(), properties, this);
-            interface->scanServices();
-
-            if (interface->type() == QNetworkSettingsType::Wired) {
-                m_interfaceModel.insert(0, interface);
-            }
-            else if (interface->type() == QNetworkSettingsType::Wifi) {
-                m_interfaceModel.append(interface);
-            }
-            emit q->interfacesChanged();
+        ConnmanSettingsInterface* interface = qobject_cast<ConnmanSettingsInterface*>(item);
+        if (interface->path() == technology.path()) {
+            return; // we already know the interface/technology
         }
     }
+
+    ConnmanSettingsInterface *interface = new ConnmanSettingsInterface(technology.path(), properties, this);
+    interface->scanServices();
+
+    if (interface->type() == QNetworkSettingsType::Wired) {
+        m_interfaceModel.insert(0, interface);
+    }
+    else if (interface->type() == QNetworkSettingsType::Wifi) {
+        m_interfaceModel.append(interface);
+    }
+
+    emit q->interfacesChanged();
 }
 
 void QNetworkSettingsManagerPrivate::onTechnologyRemoved(const QDBusObjectPath &technology)
@@ -257,9 +260,21 @@ void QNetworkSettingsManagerPrivate::onServicesChanged(ConnmanMapStructList chan
             emit q->servicesChanged();
 
         if (auto service = m_unnamedServices.take(path))
-            service->deleteLater();
+            if (!service->placeholderState()) {
+                disconnect(service, &QNetworkSettingsService::nameChanged,
+                           this, &QNetworkSettingsManagerPrivate::serviceReady);
+                disconnect(service, &QNetworkSettingsService::typeChanged,
+                           this, &QNetworkSettingsManagerPrivate::serviceReady);
+                service->deleteLater();
+            }
         if (auto service = m_unknownServices.take(path))
-            service->deleteLater();
+            if (!service->placeholderState()) {
+                disconnect(service, &QNetworkSettingsService::nameChanged,
+                           this, &QNetworkSettingsManagerPrivate::serviceReady);
+                disconnect(service, &QNetworkSettingsService::typeChanged,
+                           this, &QNetworkSettingsManagerPrivate::serviceReady);
+                service->deleteLater();
+            }
         m_unnamedServicesForSsidConnection.remove(path); // do not delete here
     }
